--- src/features/problem/retrieveDetail/SolveProblems/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './SolveProblemsPage'
