+++ src/features/problem/retrieveDetail/index.tsx
